Add files via upload
This version handles the major frequency ranges of a Heathkit HW-100/101 or SB-100/101/102, as well as allowing WWV at 15 MHz and CHU at several frequencies to be received.   It expects a 10K one-turn potentiometer.   The output of the Raduino (red wire, brown=ground) are used instead of the transceiver's VFO output.    

For maximum accuracy, you would want to measure
1)  the actual frequency of your Raduino's crystal (using a program that causes the raduino to output 25 MHz)
2)  the frequency of each of the heterodyne crystals in the heathkit
3)  the frequency of each of the BFO crystals in the heathkit.

2 & 3 can be done using an accurate receiver and a small pickup wire anywhere near the oscillators inside the Heathkit.